LayoutTests/platform/mac-wk1/http/tests/security/mixedContent/secure-redirect-to-secure-redirect-to-basic-auth-insecure-image-UpgradeMixedContent.https-expected.txt

CONSOLE MESSAGE: Blocked https://localhost:8443/resources/redirect.py?url=http://localhost:8080/security/mixedContent/resources/subresource/protected-image.py from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/mixedContent/secure-redirect-to-secure-redirect-to-basic-auth-insecure-image-UpgradeMixedContent.https.html requested insecure content from http://localhost:8080/security/mixedContent/resources/subresource/protected-image.py. This content was automatically upgraded and should be served over HTTPS.

This test loads a secure image that redirects to an secure image that redirects to an insecure image guarded by basic authentication. The insecure image should be blocked because it requires credentials.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS did not load image.
PASS successfullyParsed is true

TEST COMPLETE
